+++ lib/pages/chat_list/chat_list.dart
@@ -1,3 +1,10 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (c) 2021-2026 FluffyChat Contributors
+// Copyright (c) 2026 Simon
+//
+// MODIFICATIONS:
+// - 2026-02-05: Route and filter story rooms by displayname prefix - Simon
+
 import 'dart:async';
 
 import 'package:flutter/material.dart';
@@ -15,6 +22,7 @@ import 'package:fluffychat/l10n/l10n.dart';
 import 'package:fluffychat/pages/chat_list/chat_list_view.dart';
 import 'package:fluffychat/utils/localized_exception_extension.dart';
 import 'package:fluffychat/utils/matrix_sdk_extensions/matrix_locals.dart';
+import 'package:fluffychat/utils/story_room_extension.dart';
 import 'package:fluffychat/utils/platform_infos.dart';
 import 'package:fluffychat/utils/show_scaffold_dialog.dart';
 import 'package:fluffychat/utils/show_update_snackbar.dart';
@@ -38,7 +46,7 @@ enum PopupMenuAction {
   archive,
 }
 
-enum ActiveFilter { allChats, messages, groups, unread, spaces }
+enum ActiveFilter { allChats, messages, groups, unread, spaces, stories }
 
 extension LocalizedActiveFilter on ActiveFilter {
   String toLocalizedString(BuildContext context) {
@@ -53,6 +61,8 @@ extension LocalizedActiveFilter on ActiveFilter {
         return L10n.of(context).groups;
       case ActiveFilter.spaces:
         return L10n.of(context).spaces;
+      case ActiveFilter.stories:
+        return L10n.of(context).stories;
     }
   }
 }
@@ -133,27 +143,44 @@ class ChatListController extends State<ChatList>
       return;
     }
 
+    if (room.isStory) {
+      context.go('/rooms/story/${room.id}');
+      return;
+    }
+
     context.go('/rooms/${room.id}');
   }
 
   bool Function(Room) getRoomFilterByActiveFilter(ActiveFilter activeFilter) {
     switch (activeFilter) {
       case ActiveFilter.allChats:
-        return (room) => true;
+        return (room) => !room.isStory;
       case ActiveFilter.messages:
-        return (room) => !room.isSpace && room.isDirectChat;
+        return (room) => !room.isStory && !room.isSpace && room.isDirectChat;
       case ActiveFilter.groups:
-        return (room) => !room.isSpace && !room.isDirectChat;
+        return (room) => !room.isStory && !room.isSpace && !room.isDirectChat;
       case ActiveFilter.unread:
-        return (room) => room.isUnreadOrInvited;
+        return (room) => !room.isStory && room.isUnreadOrInvited;
       case ActiveFilter.spaces:
-        return (room) => room.isSpace;
+        return (room) => !room.isStory && room.isSpace;
+      case ActiveFilter.stories:
+        return (room) => room.isStory;
     }
   }
 
   List<Room> get filteredRooms => Matrix.of(
     context,
   ).client.rooms.where(getRoomFilterByActiveFilter(activeFilter)).toList();
+
+  List<Room> get storyRooms {
+    final rooms = Matrix.of(
+      context,
+    ).client.rooms.where((room) => room.isStory).toList();
+    rooms.sort(
+      (a, b) => b.latestEventReceivedTime.compareTo(a.latestEventReceivedTime),
+    );
+    return rooms;
+  }
 
   bool isSearchMode = false;
   Future<QueryPublicRoomsResponse>? publicRoomsResponse;